Version 4
This is actually version 3 of Creating an input system for our coffee.  But it says version 4 because the upload before this one (version 3) was supposed to be version 2. This was a result of a typo.  So this is version 4, just to make it simpler to follow along with the previous version (that was a typo).