v1.2
Something went wrong with all the uplaods of this file, so I restarted it to Version 1, I think I mentioned this in the previous upload. SO basically I didn't update visual studio code and it gave me problems, so now I just uploaded it. But I think this is my fifth time editing it. (Added comments to this one to make to code more understandable, for me at least.)